Replace workflow_presenter with override, add test for ovevrridden actions method
